function_overloading

+++ 7OOPS_2/4_overloading/4.1_function_overloading/3.cpp
@@ -7,6 +7,31 @@
 
 // question 1 - this gives error-
 
+// #include<iostream>  
+// using namespace std;  
+// void fun(int);  
+// void fun(float);  
+// void fun(int i)  
+// {  
+//     std::cout << "Value of i is : " <<i<< std::endl;  
+// }  
+// void fun(float j)  
+// {  
+//     std::cout << "Value of j is : " <<j<< std::endl;  
+// }  
+// int main()  
+// {  
+//     fun(12);  
+
+//     fun(1.2);  
+//     return 0;  
+// } 
+
+
+
+// but this don't why??
+// question 1 part continued.
+
 #include<iostream>  
 using namespace std;  
 void fun(int);  
@@ -15,14 +40,13 @@ void fun(int i)
 {  
     std::cout << "Value of i is : " <<i<< std::endl;  
 }  
-void fun(float j)  
+void fun(double j)  
 {  
     std::cout << "Value of j is : " <<j<< std::endl;  
 }  
 int main()  
 {  
     fun(12);  
-
     fun(1.2);  
     return 0;  
 }
